requirements.txt
click==8.1.3
dominate==2.7.0
Flask==2.2.2
Flask-Bootstrap==3.3.7.1
Flask-JWT==0.3.2
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
PyJWT==1.4.2
python-dotenv==0.21.0
SQLAlchemy==1.4.46
SQLAlchemy-serializer==1.4.1
visitor==0.1.3
Werkzeug==2.2.2
zipp==3.11.0
